+++ x/exchange/client/cli/tx_test.go
@@ -12,7 +12,7 @@ import (
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
 
-	"github.com/provenance-io/provenance/internal/helpers"
+	internalcollections "github.com/provenance-io/provenance/internal/collections"
 	"github.com/provenance-io/provenance/x/exchange"
 	"github.com/provenance-io/provenance/x/exchange/client/cli"
 )
@@ -1010,7 +1010,7 @@ func (s *CmdTestSuite) TestCmdTxMarketManagePermissions() {
 					}
 				}
 
-				addrOrder := helpers.Keys(expPerms)
+				addrOrder := internalcollections.Keys(expPerms)
 				sort.Slice(addrOrder, func(i, j int) bool {
 					return bytes.Compare(s.accountAddrs[addrOrder[i]], s.accountAddrs[addrOrder[j]]) < 0
 				})
@@ -1042,7 +1042,7 @@ func (s *CmdTestSuite) TestCmdTxMarketManagePermissions() {
 					3: {exchange.Permission_cancel, exchange.Permission_attributes},
 				}
 
-				addrOrder := helpers.Keys(expPerms)
+				addrOrder := internalcollections.Keys(expPerms)
 				sort.Slice(addrOrder, func(i, j int) bool {
 					return bytes.Compare(s.accountAddrs[addrOrder[i]], s.accountAddrs[addrOrder[j]]) < 0
 				})
@@ -1760,7 +1760,7 @@ func (s *CmdTestSuite) TestCmdTxGovCloseMarket() {
 	}
 }
 
-func (s *CmdTestSuite) TestCmdTxGovUpdateParams() {
+func (s *CmdTestSuite) TestCmdTxUpdateParams() {
 	tests := []txCmdTestCase{
 		{
 			name:     "cmd error",
@@ -1780,7 +1780,7 @@ func (s *CmdTestSuite) TestCmdTxGovUpdateParams() {
 		{
 			name: "prop created",
 			preRun: func() ([]string, func(*sdk.TxResponse)) {
-				expMsg := &exchange.MsgGovUpdateParamsRequest{
+				expMsg := &exchange.MsgUpdateParamsRequest{
 					Authority: cli.AuthorityAddr.String(),
 					Params: exchange.Params{
 						DefaultSplit: 777,