+++ x/ibcratelimit/simulation/genesis.go
@@ -8,7 +8,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/types/module"
 	simtypes "github.com/cosmos/cosmos-sdk/types/simulation"
 
-	"github.com/provenance-io/provenance/internal/helpers"
+	internalrand "github.com/provenance-io/provenance/internal/rand"
 	"github.com/provenance-io/provenance/x/ibcratelimit"
 )
 
@@ -19,7 +19,7 @@ const (
 
 // ContractFn randomized contract address
 func ContractFn(r *rand.Rand, accs []simtypes.Account) string {
-	randomAccount, _ := helpers.SelectRandomAccounts(r, accs, 1)
+	randomAccount, _ := internalrand.SelectAccounts(r, accs, 1)
 	if r.Intn(2) > 0 || len(randomAccount) == 0 {
 		return ""
 	}